2003-11-11 12:19 UTC-0800 Ron Pinkas <ron@ronpinkas.com>   * include/classes.h   * include/hbapi.h   * include/inet.h   * include/simplex.c   * source/codepage/cdpltwin.c   * source/lang/msgltwin.c   * source/macro/macro.slx   * source/pp/ppcore.c   * source/rtl/filesys.c   * source/rtl/fssize.c   * source/rtl/fstemp.c   * source/rtl/gtapi.c   * source/rtl/isprint.c   * source/rtl/philes.c   * source/rtl/win32ole.prg   * source/rtl/gtwin/gtwin.c   * source/vm/classes.c   * source/vm/errorapi.c   * source/vm/eval.c   * source/vm/hash.c   * source/vm/inet.c   * source/vm/runner.c   * source/vm/thread.c     * Fixed compilation warnings.

+++ xharbour/source/rtl/fssize.c
@@ -1,5 +1,5 @@
 /*
- * $Id: fssize.c,v 1.9 2001/05/15 13:02:06 vszakats Exp $
+ * $Id: fssize.c,v 1.1.1.1 2001/12/21 10:41:35 ronpinkas Exp $
  */
 
 /*
@@ -95,7 +95,7 @@ ULONG hb_fsFSize( BYTE * pszFileName, BOOL bUseDirEntry )
       }
    }
 
-   hb_fsSetError( FS_ERROR );
+   hb_fsSetError( (unsigned short) FS_ERROR );
    return 0;
 }
